Some test changes
